+++ changelog.md
@@ -3,7 +3,11 @@
 ## Upcoming
 * Cost for repairing a tool now scales with the percentage durability repaired
   * E.g. repairing a tool for half it's durability will cost only half levels
-* Books can no longer be enchanted in an ensorceller
+* Fixed the anvil running vanilla behaviour when leaving the mixin
+* Ensorceller
+  * Items enchanted in an ensorceller can no longer be merged
+  * Increased max pool from 12 to 15 (16 to 20 for full pool)
+  * Books can no longer be enchanted in an ensorceller
 
 ## 0.14.1
 * Setting your spawn point on a respawn Obelisk now saves the previous one